requirements.txt
Sphinx
flake8
ipython
mypy
numpydoc
pandas-stubs
pylint
pyls-isort
pyls-memestra
pylsp-mypy
pylsp-rope
pytest
pytest-cov
python-lsp-server[all]
types-PyYAML
